+++ include/solvers/vof.h
@@ -144,6 +144,18 @@ public:
       this->triangulation,
       this->simulation_control,
       this->multiphysics);
+
+    if (simulation_parameters.multiphysics.vof_parameters
+          .geometric_interface_reinitialization.enable)
+      {
+        this->signed_distance_solver = std::make_shared<
+          InterfaceTools::SignedDistanceSolver<dim, GlobalVectorType>>(
+          triangulation,
+          fe,
+          simulation_parameters.multiphysics.vof_parameters
+            .geometric_interface_reinitialization.max_reinitialization_distance,
+          0.0);
+      }
   }
 
   /**
@@ -426,7 +438,6 @@ public:
     return this->subequations->get_solution(
       VOFSubequationsID::curvature_projection);
   }
-
   /**
    * @brief Output the L2 and Linfty norms of the correction vector.
    *
@@ -674,6 +685,13 @@ private:
   void
   reinitialize_interface_with_algebraic_method();
 
+  /**
+   * @brief Reinitialize the interface between fluids using the geometric
+   * approach.
+   */
+  void
+  reinitialize_interface_with_geometric_method();
+
   GlobalVectorType nodal_phase_fraction_owned;
 
   MultiphysicsInterface<dim> *multiphysics;
@@ -750,6 +768,10 @@ private:
 
   // Phase fraction filter
   std::shared_ptr<VolumeOfFluidFilterBase> filter;
+
+  // Signed distance solver for geometric redistanciation
+  std::shared_ptr<InterfaceTools::SignedDistanceSolver<dim, GlobalVectorType>>
+    signed_distance_solver;
 };